Add colours for patch-style diffs
When vim displays patches (like in Gundo or Gitv), it uses a different
set of highlight groups that have no syntax highlighting. Use something
closer to jellybeans.

It blends in better with sandydune if I flip the fg and bg colours, but
I think I like the classic jellybeans look better. Will have to try it
for a bit.