Use a nicer colour for search: blue/black
'brown' is pretty ugly. Let's take inspiration from the new QuickFixLine
and use SkyBlue. Change QuickFixLine to be lighter so inline search
matches are somewhat visible.

This gentler Search colour is easier to deal with many search results.
The IncSearch colour is bright enough to get attention, but dimmed a bit
to be easy on the eyes. Dimmed each colour channel by 0x20.